parkinglot -> vacancy
