--- src/pages/parkinglot/Parkinglot.module.scss
@@ -1,33 +1,0 @@
-.root {
-}
-
-.mainTableWidget {
-  position: relative;
-  width: 800px;
-  height: 600px;
-}
-
-.small {
-  font-size: xx-small;
-}
-
-.slot {
-  border: 2px solid gray;
-  background-color: red;
-  position: absolute;
-  &.horizontal {
-    transform: rotate(90deg);
-  }
-  &.vertical {
-    transform: rotate(0deg);
-  }
-  &.leftTilt {
-    transform: rotate(315deg);
-  }
-  &.rightTilt {
-    transform: rotate(45deg);
-  }
-  &.isEmpty {
-    background-color: white;
-  }
-}
