--- src/pages/parkinglot/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "parkinglot",
-  "version": "0.0.0",
-  "private": true,
-  "main": "./Parkinglot.js"
-}
