+++ src/pages/vacancy/package.json
@@ -1,5 +1,6 @@
 {
   "name": "vacancy",
-  "version": "1.0.0",
-  "main": "vacancy.js"
+  "version": "0.0.0",
+  "private": true,
+  "main": "./vacancy.js"
 }
